fixed bug in class.tx_commercecoupons_baskethooks.php
git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/commercecoupons/trunk@9415 735d13b6-9817-0410-8766-e36946ffe9aa

+++ hooks/class.tx_commercecoupons_baskethooks.php
@@ -71,9 +71,11 @@ class tx_commercecoupons_baskethooks extends tx_commerce_basic_basket {	// exten
 			$deleteAllCoupons = false;
 			foreach($article_id as $art => $aObj){
 				if(intval($aObj['count']) == 0 || intval($aObj['count']) > 1){
-					foreach($coupons as $coupon) {
-						if(isset($coupon['_addedArticles'][$art])) {
-							$deleteAllCoupons = true;
+					if(is_array($coupons)) {
+						foreach($coupons as $coupon) {
+							if(isset($coupon['_addedArticles'][$art])) {
+								$deleteAllCoupons = true;
+							}
 						}
 					}
 				}
